requirements.txt
flask
opencv-python-headless
requests
torch
transformers
Pillow
